Update type signatures for all files
- Update type signatures for utils.jl
- Update type signatures for bus.jl
- Update type signatures for capacitors and controls
- Update type signatures in circuit.jl
- Update type signatures for ckt_element.jl
- Update type signatures for ckt_element.jl
- Update type signatures for ctrl_queue.jl
- Add type signature in element, error and executive
- Add type signature in fuses, generators, isource, linecodes, lines and loads
- Add type signature in meters, monitors, parser and pdelements
- Add type signature in progress, properties and pvsystems
- Add type signature in reclosers, reg_controls and relays
- Fix reg_controls, reclosers and add type signatures for sensors
- Add type signatures for settings
- Add type signatures for solution
- Add type signatures for swt_controls
- Add type signatures in topology
- Add type signatures in transformers
- Add type signatures to vsources.jl
- Add type signatures to xycurves.jl

+++ src/progress.jl
@@ -17,7 +17,7 @@ module Progress
     end
 
     """(write-only) Percent progress to indicate [0..100]"""
-    function PctProgress(Value)
+    function PctProgress(Value::Int)
         lib.DSSProgress_Set_PctProgress(Value)
     end
